chore(CDAC): moved test-helper to __test__/helpers
